==> 4_Analysis.Rmd <==
---
title: "4_Analysis"
output: html_document
---

```{r, include = F}
# Startup
library(tidyverse)
library(ggpubr)
library(scales)

rm(list = ls())

# Load Data
df <- readRDS(file = "S:\\COPE\\Data\\Prediction\\Analysis-Ready.rds")

# Prepare Data
# Set luck threshold (in SDs)
luck_threshold = .5

# Make RTI a *positive* number - higher is *better*
df <- df %>%
  mutate(across(starts_with("rti"),
                ~ - .))

# Create new variables (RTI is already in SD units, per 1_Data Preparation.R)
df <- df %>%
  mutate(preference_pp = rti_pred_pp - rti_pred_abc,
         preference_abc = rti_pred_abc - rti_pred_pp,
         preferred_condition = case_when(
           preference_pp > luck_threshold ~ "Project Personality",
           preference_abc > luck_threshold ~ "ABC Project",
           TRUE ~ "No Preference"
         ),
         preferred_condition_force = case_when(
           preference_pp > 0 ~ "Project Personality",
           preference_abc > 0 ~ "ABC Project",
           TRUE ~ "No Preference"
         ),
         luck = case_when(
           preferred_condition != "No Preference" & preferred_condition == condition ~ "Lucky",
           preferred_condition != "No Preference" & preferred_condition != condition ~ "Unlucky",
           preferred_condition == "No Preference" ~ "No Luck"
         ),
         luck_force = case_when(
           preferred_condition_force != "No Preference" & preferred_condition_force == condition ~ "Lucky",
           preferred_condition_force != "No Preference" & preferred_condition_force != condition ~ "Unlucky",
           preferred_condition_force == "No Preference" ~ "No Luck"
         ))
```

## Descriptive Statistics

```{r, echo = F}
full_sample <- readRDS(file = "S:\\COPE\\Data\\Prediction\\Full Sample.rds")

full_sample %>%
  select(starts_with("b_dem")) %>%
  map(~ data.frame(x = .x) %>%
        count(x) %>%
        mutate(pct = percent(n / sum(n), .01)))

full_sample %>%
  summarize(b_cdi_mean_mean = mean(b_cdi_mean),
            b_cdi_mean_sd = sd(b_cdi_mean))

df %>%
  group_by(condition) %>%
  summarize(across(starts_with("rti"), list(mean = mean, sd = sd))) %>%
  t()
```

## PAI

```{r, echo = F}
df %>%
  rowwise() %>%
  mutate(preference = max(preference_pp, preference_abc)) %>%
  ungroup() %>%
  summarize(preference_mean = mean(preference),
            preference_sd = sd(preference))
```

## Inferential Statistics

```{r, echo = F}
df %>%
  group_by(luck) %>%
  summarize(n = n(),
            rti_mean = mean(rti_actual),
            rti_sd = sd(rti_actual))

df %>%
  group_by(luck_force) %>%
  summarize(n = n(),
            rti_mean = mean(rti_actual),
            rti_sd = sd(rti_actual))

## Tests
temp <- df %>%
  filter(luck %in% c("Lucky", "Unlucky"))
# t.test(temp$rti_actual ~ temp$luck)

temp <- df %>%
  filter(luck_force %in% c("Lucky", "Unlucky"))
t.test(temp$rti_actual ~ temp$luck_force)  
```

## Visualizations

```{r, echo = F}
# Comparison of outcomes
df %>%
  group_by(luck_force) %>%
  summarize(n = n(),
            rti_mean = mean(rti_actual),
            rti_sd = sd(rti_actual)) %>%
  mutate(rti_ci = 1.96 * (rti_sd / sqrt(n))) %>%
  ggplot() +
    geom_col(aes(luck_force, rti_mean)) +
    geom_errorbar(aes(luck_force, ymin = rti_mean - rti_ci, ymax = rti_mean + rti_ci), size = 2, width = .5) +
    scale_x_discrete(name = NULL) +
    scale_y_continuous(name = "RTI") +
    ggtitle("Actual RTI for Lucky and Unlucky Participants", "Means and 95% Confidence Intervals") +
    theme_classic()

# Distributions of predictions
df %>%
  pivot_longer(matches("^rti")) %>%
  ggplot(aes(value)) +
    geom_density() +
    facet_wrap( ~ name)

# Correlations between predicted and actual RTI
df %>%
  pivot_longer(matches("^rti_pred")) %>%
  filter((name == "rti_pred_pp" & condition == "Project Personality") |
         (name == "rti_pred_abc" & condition == "ABC Project")) %>%
  mutate(name = case_when(name == "rti_pred_abc" ~ "Action Brings Change Project",
                          name == "rti_pred_pp" ~ "Project Personality")) %>%
  ggplot(aes(rti_actual, value)) +
    geom_point() +
    scale_x_continuous(name = "Actual RTI", breaks = seq(-2, 4, 2)) +
    scale_y_continuous(name = "Predicted RTI") +
    stat_cor() +
    facet_wrap(~ name) +
    ggtitle("Correlations Between Predicted RTI and Actual RTI") +
    theme_classic()

# Correlation between predicted RTI across groups
df %>%
  ggplot(aes(rti_pred_abc, rti_pred_pp)) +
    geom_point() +
    scale_x_continuous(name = "Predicted RTI (ABC)") +
    scale_y_continuous(name = "Predicted RTI (PP)") +
    stat_cor() +
    ggtitle("Correlation Between Predicted RTI in Both Groups") +
    theme_classic()

# Funky plot
df %>%
  group_by(condition) %>%
  arrange(condition, -rti_actual) %>%
  mutate(participant = row_number(),
         preferred_condition = case_when(
           rti_pred_pp < rti_pred_abc ~ "Project Personality",
           rti_pred_abc < rti_pred_pp ~ "ABC Project"),
         preferred_rti = case_when(
           rti_pred_pp < rti_pred_abc ~ rti_pred_pp,
           rti_pred_abc < rti_pred_pp ~ rti_pred_abc),
         other_condition = case_when(
           rti_pred_pp < rti_pred_abc ~ "ABC Project",
           rti_pred_abc < rti_pred_pp ~ "Project Personality"),
         other_rti = case_when(
           rti_pred_pp < rti_pred_abc ~ rti_pred_abc,
           rti_pred_abc < rti_pred_pp ~ rti_pred_pp)
         ) %>%
  ungroup() %>%
  ggplot() +
    geom_point(aes(participant, rti_actual), color = "#F9B650") +
    geom_point(aes(participant, preferred_rti, color = preferred_condition)) +
    geom_point(aes(participant, other_rti, color = other_condition)) +
    geom_segment(aes(x = participant, xend = participant, 
                     y = rti_pred_abc, yend = rti_pred_pp,
                     color = preferred_condition)) +
    scale_color_discrete(name = NULL) +
    scale_x_continuous(name = NULL) +
    scale_y_reverse(name = "RTI (Lower is Better)", breaks = -4:2) +
    coord_flip() +
    facet_wrap(~ condition) +
    ggtitle("Actual and Predicted RTI, by Condition") +
    theme_minimal() +
    theme(axis.ticks.y = element_blank(),
          axis.text.y = element_blank(),
          panel.grid.major.y = element_blank(),
          panel.grid.minor.y = element_blank(),
          legend.position = "top",
          plot.title = element_text(hjust = .5))
```
